feat: add useInlineSVG hook
